% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{box_blur}
\alias{box_blur}
\title{Apply box blur to a matrix}
\usage{
box_blur(m, d = 5)
}
\arguments{
\item{m}{a matrix of values to be blurred.}

\item{d}{the distance either side of each pixel that is searched when
blurring}
}
\description{
Smooths a matrix of values by applying a box blur, in which each
  pixel of a matrix is replaced with the average value of pixels in a box
  around it.
}
