% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\name{col_plot}
\alias{col_plot}
\title{Simple plot to explore colour palettes}
\usage{
col_plot(x, size = 8)
}
\arguments{
\item{x}{colour series}

\item{size}{the size of plotting circles}
}
\description{
Simple plot to explore colour palettes.
}
