% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\name{col_tim}
\alias{col_tim}
\title{tim.colors from fields package}
\usage{
col_tim(n = 10)
}
\arguments{
\item{n}{the number of colours}
}
\description{
Blue-to-red colours. Full credit to tim.colors from the fields package, from
}
