% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{dist_gc}
\alias{dist_gc}
\title{Calculate pairwise great circle distance between points}
\usage{
dist_gc(x)
}
\arguments{
\item{x}{a two-column matrix or dataframe with longitude in the first column
and latitude in the second.}
}
\description{
Analogue of the \code{dist()} function, but calculating great
  circle distances (distance on the surface of a sphere). Points are input as
  longitude and latitude.
}
