% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{dna_complement}
\alias{dna_complement}
\title{Return complementary DNA or RNA sequence}
\usage{
dna_complement(x, format_rna = FALSE)
}
\arguments{
\item{x}{input DNA sequence as character string.}

\item{format_rna}{logical. If \code{TRUE} then A complements to U, rather than T.}
}
\description{
Return complementary DNA or RNA sequence.
}
