% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{dna_to_aa}
\alias{dna_to_aa}
\title{Convert DNA sequence to amino-acid}
\usage{
dna_to_aa(x, output_format = 1)
}
\arguments{
\item{x}{three-letter character sequence of A, T, G and C.}

\item{output_format}{1 for full name, 2 for single-letter code.}
}
\description{
Convert three-letter sequence of ATGC to corrseponding
  translated amino-adid sequence. Amino-acids can be reported in full name or
  single-letter code. Function is most efficient when using a vector of
  sequences.
}
