% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{fractal_noise}
\alias{fractal_noise}
\title{Generate fractal noise}
\usage{
fractal_noise(out_rows = 128, out_cols = 128, a = 0.6,
  n_levels = 7)
}
\arguments{
\item{out_rows, out_cols}{rows and columnds in output matrix.}

\item{a}{scale parameter. Values near zero place more weight on large
features, a value of 1 places equal weight on all levels, and values > 1
place more weight on small features.}

\item{n_levels}{number of levels of Perlin noise to layer.}
}
\description{
Generates 2D fractal noise by layering Perlin noise at different levels.
}
