% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{func_header}
\alias{func_header}
\title{Produce function header text}
\usage{
func_header(type = 1)
}
\arguments{
\item{type}{switches between the following text blocks:
\itemize{
  \item{type = 1: basic header}
  \item{type = 2: complete header with all bells and whistles}
  \item{type = 3: not included header}
}}
}
\description{
Produce header text for a single function to paste into script
  (automatically copied to the clipboard). Header text is written in roxygen
  format.
}
