% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{gamma_incomplete}
\alias{gamma_incomplete}
\title{The Incomplete Gamma Function}
\usage{
gamma_incomplete(s, x)
}
\arguments{
\item{s}{shape parameter}

\item{x}{lower bound of integral}
}
\description{
Returns the value of the incomplete gamma function, as defined
  \href{https://en.wikipedia.org/wiki/Incomplete_gamma_function#Definition}{here}.
  Accepts vector inputs.
}
