% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_projection}
\alias{get_projection}
\title{Get projection matrix}
\usage{
get_projection(x_lim, y_lim, z_lim, theta, phi, d = 2)
}
\arguments{
\item{x_lim, y_lim, z_lim}{limits of \code{persp()} plot.}

\item{theta}{angle of rotation (degrees).}

\item{phi}{angle of elevation (degrees).}

\item{d}{strength of perspective transformation.}
}
\description{
Get projection matrix from \code{persp()} function without
  producing plot. The angles of projection are given by \code{theta}
  (rotation) and \code{phi} (elevation), and the strength of perspective
  transformation is given by \code{d}.
}
