% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{gg3d_add_grid_lines}
\alias{gg3d_add_grid_lines}
\title{Add grid lines to ggplot perspective plot}
\usage{
gg3d_add_grid_lines(
  myplot,
  x = seq(0, 1, 0.1),
  y = c(0, 1),
  z = 0,
  proj_mat,
  col = grey(0.8),
  size = 0.5,
  break_axis = NULL
)
}
\arguments{
\item{myplot}{an object of class \code{ggplot}.}

\item{x, y, z}{coordinates of lines. Values must be a sequence of breaks
\emph{or} a pair of start-end values \emph{or} a single value (see
details).}

\item{proj_mat}{4*4 projection matrix, as returned from
\code{get_projection()}.}

\item{col}{line colour.}

\item{size}{line size.}

\item{break_axis}{which axis to apply breaks over. If \code{NULL} then chosen
automatically from other inputs, otherwise must be one of \code{"x"},
\code{"y"} or \code{"z"} (see details).}
}
\description{
Add grid lines to a ggplot perspective plot.
}
\details{
The vectors \code{x}, \code{y}, and \code{z} together define the
  locations and orientations of the lines. One of these vectors must be a
  series of breaks, one must be a limit (i.e. a vector of two values) and one
  must be a single value. For example, if \code{x = 1:5, y = c(-1,1), z = 3}
  then lines will be drawn parallel to the y-axis at x-values 1:5, spanning a
  range -1 to 1, and in the z-plane at position z = 3. If there are only two
  breaks then it becomes impossible to determine which axis represents breaks
  and which represents limits, hence the argument \code{break_axis} must be
  specified (otherwise this is chosen automatically).
}
