% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{gg3d_scatterplot}
\alias{gg3d_scatterplot}
\title{Produce a 3D scatterplot in ggplot format}
\usage{
gg3d_scatterplot(x, y, z, colour = 1, size = 0.5, theta = 135,
  phi = 30, d = 2, x_lim = NULL, y_lim = NULL, z_lim = NULL,
  x_grid = NULL, y_grid = NULL, z_grid = NULL, z_type = 2,
  flip_grid_x = FALSE, flip_grid_y = FALSE, grid_col = grey(0.8),
  grid_size = 0.25, axis_col = "black", axis_size = 0.5,
  x_lab = "x", y_lab = "y", z_lab = "z", tick_length = 0.2,
  axis_lab_size = 3, axis_lab_dist = 2)
}
\arguments{
\item{x}{vector of values in the x-dimension.}

\item{y}{vector of values in the y-dimension.}

\item{z}{vector of values in the z-dimension.}

\item{colour}{vector specifying point colours. Can be continuous or discrete.}

\item{size}{size of data points.}

\item{theta}{angle of rotation (degrees).}

\item{phi}{angle of elevation (degrees).}

\item{d}{strength of perspective transformation.}

\item{x_lim}{plotting limits in the x-dimension.}

\item{y_lim}{plotting limits in the y-dimension.}

\item{z_lim}{plotting limits in the z-dimension.}

\item{x_grid}{sequence of x-breaks defining grid.}

\item{y_grid}{sequence of y-breaks defining grid.}

\item{z_grid}{sequence of z-breaks defining grid.}

\item{z_type}{switch between horizontal grid in the z-dimension (\code{z_type
= 1}) and both horizontal and vertical grid (\code{z_type = 2}).}

\item{flip_grid_x}{if \code{TRUE} then the vertical grid in the x-axis is
moved to the other side of the plot.}

\item{flip_grid_y}{if \code{TRUE} then the vertical grid in the y-axis is
moved to the other side of the plot.}

\item{grid_col}{the colour of grid lines.}

\item{grid_size}{the size of grid lines.}

\item{axis_col}{the colour of axis lines.}

\item{axis_size}{the size of axis lines.}

\item{x_lab}{the x-axis label.}

\item{y_lab}{the y-axis label.}

\item{z_lab}{the z-axis label.}

\item{tick_length}{the absolute length of axis ticks.}

\item{axis_lab_size}{the size of axis labels.}

\item{axis_lab_dist}{the absolute distance of axis labels from the edge of
the plotting region.}
}
\description{
Produce a 3D scatterplot in ggplot format.
}
