% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{head_mat}
\alias{head_mat}
\title{Head function for matrices}
\usage{
head_mat(x, n, m = n)
}
\arguments{
\item{x}{a matrix.}

\item{n}{the number of rows to print.}

\item{m}{the number of cols to print (default m = n).}
}
\description{
The \code{head()} function is annoying when applied to a matrix,
  as it prints just the first n rows, which doesn't work well for wide
  matrices. This function prints just the first n rows and m columns (n = m
  by default).
}
