% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{header}
\alias{header}
\title{Produce script header text}
\usage{
header(type = 1)
}
\arguments{
\item{type}{switches between the following text blocks:
\itemize{
  \item{type = 1: basic header including name, author, date and purpose}
  \item{type = 2: as above, plus list of useful commands and shortcuts when
  working with packages}
}}
}
\description{
Produce header text to paste into new script (automatically 
  copied to the clipboard). Different header types are available for
  different types of script (see below).
}
