% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{invgamma}
\alias{invgamma}
\alias{dinvgamma}
\alias{pinvgamma}
\alias{qinvgamma}
\alias{rinvgamma}
\title{The Inverse-Gamma Distribution}
\usage{
dinvgamma(x, alpha, beta, log_on = FALSE)

pinvgamma(q, alpha, beta, lower_tail = TRUE)

qinvgamma(p, alpha, beta, lower_tail = TRUE)

rinvgamma(n, alpha, beta)
}
\arguments{
\item{x}{vector of values or quantiles.}

\item{alpha}{shape parameter.}

\item{beta}{scale parameter.}

\item{log_on}{logical; if \code{TRUE} values are returned logged.}

\item{q}{vector of quantiles}

\item{lower_tail}{logical; if \code{TRUE} (default) probabilities are
\emph{P[X <= x]} otherwise \emph{P[X > x]}.}

\item{p}{vector of probabilities.}

\item{n}{number of observations.}
}
\description{
Density, distribution function, quantile function and random
  generation for the inverse-gamma distribution with shape \code{alpha} and
  scale \code{beta}.
}
