% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{layout_mat}
\alias{layout_mat}
\title{Create layout matrix automatically from plot number}
\usage{
layout_mat(n)
}
\arguments{
\item{n}{number of plots in final layout}
}
\description{
Create a layout matrix that can be used with gridExtra to 
  arrange ggplot objects. The dimensions of the matrix are chosen 
  automatically from the number of plots to create a rectangular arrangement
  (portrait) of sufficient size.
}
