% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{list_to_matrix}
\alias{list_to_matrix}
\title{rbind a list of matrices into a single matrix}
\usage{
list_to_matrix(l)
}
\arguments{
\item{l}{list of matrices.}
}
\description{
rbind a list of matrices into a single matrix. All matrices must
  have the same number of columns.
}
