% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{midpoints}
\alias{midpoints}
\title{Find midpoints of a vector of breaks}
\usage{
midpoints(x)
}
\arguments{
\item{x}{a vector of breaks}
}
\description{
Find midpoints of a vector of breaks. The resulting vector will
  have one less element than the input vector.
}
