% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{more_colours}
\alias{more_colours}
\title{Expand series of colours by interpolation}
\usage{
more_colours(n, raw_cols = brewer.pal(10, "Paired"))
}
\arguments{
\item{n}{how many colours to return}

\item{raw_cols}{vector of colours to interpolate}
}
\description{
Expand series of colours by interpolating a given vector of raw
  colours. The pattern of interpolation is designed so that (n+1)th
  interpolation contains the nth interpolation plus one more colour, rather
  than being a completely different series.
}
\details{
Takes a vector of values, returns the square.
}
