% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{moving_average}
\alias{moving_average}
\title{Calculate moving average over a window}
\usage{
moving_average(x, d = 3, include_ends = TRUE)
}
\arguments{
\item{x}{vector of numeric values.}

\item{d}{how far to look either side of central value. For example,
\code{d=2} would search -2:2.}

\item{include_ends}{if \code{TRUE}, values where the window extends beyond
the range of the sequence are included in the output. Otherwise return NA
for these values.}
}
\description{
Calculate moving average of a sequence of values over a defined
  window that extends \code{d} units either side of the central value (hence
  \code{d=1} returns the original sequence). NA values are not counted in the
  numerator or denominator of the average. Where the window goes beyond the
  limits of the sequence, these values can included in the moving average
  (simply ignoring values outside the range) or returned as NA.
}
