% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{perlin_noise}
\alias{perlin_noise}
\title{Generate Perlin noise}
\usage{
perlin_noise(out_rows = 128, out_cols = 128, levels_x = 10, levels_y = 10)
}
\arguments{
\item{out_rows, out_cols}{rows and columnds in output matrix.}

\item{levels_x, levels_y}{bumpyness in x and y dimension.}
}
\description{
Generates 2D Perlin noise of any scale, in a matrix of any size.
  Credit to
  \href{https://stackoverflow.com/users/1129973/vincent-zoonekynd}{Vincent
  Zoonekynd} for
  \href{https://stackoverflow.com/questions/15387328/realistic-simulated-elevation-data-in-r-perlin-noise}{this
  answer}.
}
