% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{plot_density}
\alias{plot_density}
\title{Quick plot of density distribution}
\usage{
plot_density(dist, qrange = c(0.01, 0.99), ...)
}
\arguments{
\item{dist}{the name of the distribution family. For exapmle \code{"norm"}
can be used to explore the \code{dnorm()} density.}

\item{qrange}{the quantile range over which to plot.}

\item{...}{further parameters to the distribution function.}
}
\description{
Produces simple plot of probability density for the main
  families of distributions, including normal, lognormal, gamma and
  inverse-gamma.
}
