% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{project_2d}
\alias{project_2d}
\title{Project 3D coordinates to 2D}
\usage{
project_2d(x, y, z, proj_mat)
}
\arguments{
\item{x, y, z}{world coordinates.}

\item{proj_mat}{4*4 projection matrix, as returned from
\code{get_projection()}.}
}
\description{
Use a projection matrix to transform 3D world coordinates to 2D
  coordinates - for example representing a camera viewing a 3D scene. See
  \code{get_projection()} for how to obtain a projection mtarix. This
  function is equivalent to the \code{trans3d()} function, but also stores
  depth information relative to the camera.
}
