% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{rCRP}
\alias{rCRP}
\title{Draw from Chinese restaurant process}
\usage{
rCRP(n, alpha = 1)
}
\arguments{
\item{n}{the number of random draws.}

\item{alpha}{the concentration parameter of the process.}
}
\description{
Draw a partition of n values from the Chinese restaurant process
  with concentration parameter \code{alpha}.
}
