% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{rDPM}
\alias{rDPM}
\title{Draw from Dirichlet process mixture model}
\usage{
rDPM(n, alpha = 1, d = 1, tau = 10, sigma = 1)
}
\arguments{
\item{n}{the number of random draws.}

\item{alpha}{the concentration parameter of the process.}

\item{d}{the number of dimensions.}

\item{tau}{the standard deviation of the symmetric normal prior on component
means.}

\item{sigma}{the standard deviation of the symmetric normal mixture
distribution. Equal for all components.}
}
\description{
Draw a series of point from the d-dimensional Dirichlet process
  with normal prior and normal mixture components.
}
