% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{set_compare}
\alias{set_compare}
\title{Print summary of setdiff and intersections of two sets}
\usage{
set_compare(s1, s2)
}
\arguments{
\item{s1}{first set.}

\item{s2}{second set.}
}
\description{
Given two sets s1 and s2, print three values:
  \enumerate{
    \item the number of s1 not in s2
    \item the intersection size of s1 and s2
    \item the number of s2 not in s1
  }
}
