% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{smooth_cols}
\alias{smooth_cols}
\title{Produce smooth colours from values}
\usage{
smooth_cols(x, xmin = min(x, na.rm = T), xmax = max(x, na.rm = T),
  n_levels = 1000, raw_cols = col_tim())
}
\arguments{
\item{x}{values from which to obtain colours.}

\item{xmin, xmax}{minimum and maximum values of x.}

\item{n_levels}{number of colours in palette.}

\item{raw_cols}{colours that make up the palette.}
}
\description{
Read in continuous values between xmin and xmax and return
  colours associated with these values, taken from a smoothly varying scale.
}
