% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{vec2mat}
\alias{vec2mat}
\title{Convert vector to matrix}
\usage{
vec2mat(x, y, dim)
}
\arguments{
\item{x, y}{two vectors of any type.}

\item{dim}{which dimension of the output matrix to return.}
}
\description{
Given two input vectors, expands these vectors into matrices
  based on the dimensions of the other vector. Returns either the matrix in
  x, or the matrix in y.
}
