% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bin_2d}
\alias{bin_2d}
\title{Bin values in two dimensions}
\usage{
bin_2d(x, y, x_breaks = NULL, y_breaks = NULL)
}
\arguments{
\item{x}{first dimension of values to bin}

\item{y}{second dimension of values to bin}

\item{x_breaks}{set of breaks in x-dimension}

\item{y_breaks}{set of breaks in y-dimension}
}
\description{
Bin values in two dimensions based on a vector of breaks in each
  dimension. Values are included in a bin if they are >= the left break and <
  the right break. Default breaks are chosen to encompass all the data.
}
