% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{func_header}
\alias{func_header}
\title{Produce function header text}
\usage{
func_header(type = 1)
}
\arguments{
\item{type}{switches between the following text blocks:
\enumerate{
  \item basic function header
  \item complete function header with all the bells and whistles
  \item function not exported
  \item S3 method, e.g. \code{print.my_class()}
}}
}
\description{
Produce header text for a single function to paste into script
  (automatically copied to the clipboard). Header text is written in roxygen
  format.
}
