% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{gg_inset}
\alias{gg_inset}
\title{Add inset plot to faceted ggplot}
\usage{
gg_inset(grob, xmin = -Inf, xmax = Inf, ymin = -Inf, ymax = Inf, data)
}
\arguments{
\item{grob}{a ggplot object.}

\item{xmin}{inset starts at this coordinate. Set to \code{-Inf} to always
start from far left edge.}

\item{xmax}{inset ends at this coordinate. Set to \code{Inf} to always start
from far right edge.}

\item{ymin}{inset starts at this coordinate. Set to \code{-Inf} to always
start from far bottom edge.}

\item{ymax}{inset ends at this coordinate. Set to \code{Inf} to always start
from far top edge.}

\item{data}{dataframe, should specify the value of the variable with which
the original plot is facetted (see examples).}
}
\description{
Ordinarily ggplot is not setup to add inset plots (annotations) to faceted plots, because it wants to add the same inset to every facet. This custom annotation function makes it possible to add an inset to a single facet. Credit goes to \href{https://stackoverflow.com/users/471093/baptiste}{baptiste} for \href{https://stackoverflow.com/questions/37867758/insetting-on-facet-grided-and-grid-arrangeed-plot}{this} solution.
}
\examples{
# produce main plot and inset plot
plot_df <- data.frame(x = rnorm(1e3), y = rnorm(1e3), group = 1:2)
plot_main <- ggplot2::ggplot() +
             ggplot2::geom_point(ggplot2::aes(x = x, y = y), data = plot_df) +
             ggplot2::facet_wrap(~group)
plot_inset <- ggplot2::qplot(rnorm(1e3))

# use gg_inset to annotate first panel only
plot_combined <- plot_main + gg_inset(ggplot2::ggplotGrob(plot_inset), data = data.frame(group = 1),
                                      xmin = -Inf, xmax = 1.5, ymin = 0.6, ymax = Inf)

plot_combined
}
