% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{header}
\alias{header}
\title{Produce script header text}
\usage{
header(type = 1, copy_clipboard = FALSE)
}
\arguments{
\item{type}{switches between the following text blocks:
\itemize{
  \item{type = 1: basic header including name, author, date and purpose}
  \item{type = 2: as above, plus list of useful commands and shortcuts when
  working with packages}
}}

\item{copy_clipboard}{whether the header text should also be copied to the
clipboard. This should be \code{FALSE} when running on PC as it can cause R
to crash.}
}
\description{
Produce header text to copy and paste into new script. Different
  header types are available for different types of script (see below).
}
