% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{project_2d_line}
\alias{project_2d_line}
\title{Project 3D line coordinates to 2D}
\usage{
project_2d_line(x0, y0, z0, x1, y1, z1, proj_mat)
}
\arguments{
\item{x0, y0, z0}{coordinates of start point.}

\item{x1, y1, z1}{coordinates of end point.}

\item{proj_mat}{4*4 projection matrix, as returned from
\code{get_projection()}.}
}
\description{
Use \code{project2D} to project start and end points from 3D
  world coordinates to 2D screen coordinates.
}
