% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{sim_wrightfisher}
\alias{sim_wrightfisher}
\title{Simulate allele frequencies from Wright-Fisher model}
\usage{
sim_wrightfisher(N, L, alleles, mu, mig_mat, t_out)
}
\arguments{
\item{N}{number of individuals per deme. Must be the same for all
demes.}

\item{L}{number of loci (assumed independent).}

\item{alleles}{number of alleles. Can be a single number for all loci or a
vector of length \code{L}.}

\item{mu}{mutation rate. Assumes finite-alleles model, with equal chance of
mutating from any allele to any other.}

\item{mig_mat}{migration matrix specifying the per-generation probability of
an individual migrating from any deme (in rows) to any other deme (in
columns).}

\item{t_out}{vector of times at which results will be output.}
}
\description{
Simulate Wright-Fisher evolution in a series of partially
  connected demes.
}
\details{
Assumes a haploid population and independent loci (no linkage
  disequilibrium). Implements a finite-alleles mutation model with equal
  chance of mutating from any allele to any other. Initialises allele
  frequencies by drawing from a symmetric Dirichlet(theta/k) distribution,
  where \eqn{theta = 2*N*mu}, which is the analytical equilibrium
  distribution between drift and finite-alleles mutation, ignoring migration
  between demes. Migration is implemented by proposing random swaps of
  individuals between demes, thereby ensuring population sizes remain
  constant over time. For this reason, \code{N} must be the same for all
  demes.
}
